Finalcode

+++ lib/mvc/view/familyList/famList.dart
@@ -8,7 +8,6 @@ import 'package:famlynk_version1/services/familySevice/famlist_servicess.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-
 class FamilyList extends StatefulWidget {
   @override
   _FamilyListState createState() => _FamilyListState();
@@ -44,8 +43,6 @@ class _FamilyListState extends State<FamilyList> {
       }
     }
   }
-
- 
 
   String capitalizeFirstLetter(String text) {
     if (text.isEmpty) return text;
@@ -113,71 +110,74 @@ class _FamilyListState extends State<FamilyList> {
                                   ),
                                   Spacer(),
                                   Center(
-                                    child: PopupMenuButton(
-                                      itemBuilder: (BuildContext context) => [
-                                        if (isEditable)
-                                          PopupMenuItem(
-                                            child: TextButton(
-                                              onPressed: () {
-                                                Navigator.push(
-                                                  context,
-                                                  MaterialPageRoute(
-                                                    builder: (context) =>
-                                                        UpdateFamList(
-                                                      updateMember:
-                                                          familyList[index],
+                                    child: InkWell(
+                                      child: PopupMenuButton(
+                                        itemBuilder: (BuildContext context) => [
+                                          if (isEditable)
+                                            PopupMenuItem(
+                                              child: TextButton(
+                                                onPressed: () {
+                                                  Navigator.push(
+                                                    context,
+                                                    MaterialPageRoute(
+                                                      builder: (context) =>
+                                                          UpdateFamList(
+                                                        updateMember:
+                                                            familyList[index],
+                                                      ),
                                                     ),
+                                                  );
+                                                },
+                                                child: Text(
+                                                  "Edit",
+                                                  style: TextStyle(
+                                                      color: Colors.blue),
+                                                ),
+                                              ),
+                                            ),
+                                          if (!isEditable)
+                                            PopupMenuItem(
+                                              child: TextButton(
+                                                onPressed: () {
+                                                  showMyDialog(
+                                                    familyList[index]
+                                                        .userId
+                                                        .toString(),
+                                                    familyList[index]
+                                                        .uniqueUserID
+                                                        .toString(),
+                                                  );
+                                                },
+                                                child: Text(
+                                                  "remove",
+                                                  style: TextStyle(
+                                                    color: Colors.red,
                                                   ),
-                                                );
-                                              },
-                                              child: Text(
-                                                "Edit",
-                                                style: TextStyle(
-                                                    color: Colors.blue),
+                                                ),
                                               ),
                                             ),
-                                          ),
-                                        if (!isEditable)
-                                          PopupMenuItem(
-                                            child: TextButton(
-                                              onPressed: () {
-                                                _showMyDialog(
-                                                  familyList[index]
-                                                      .userId
-                                                      .toString(),
-                                                  familyList[index]
-                                                      .uniqueUserID
-                                                      .toString(),
-                                                );
-                                              },
-                                              child: Text(
-                                                "remove",
-                                                style: TextStyle(
-                                                    color: Colors.red),
+                                          if (isEditable)
+                                            PopupMenuItem(
+                                              child: TextButton(
+                                                onPressed: () {
+                                                  _showMyDialog(
+                                                    familyList[index]
+                                                        .userId
+                                                        .toString(),
+                                                    familyList[index]
+                                                        .uniqueUserID
+                                                        .toString(),
+                                                  );
+                                                },
+                                                child: Text(
+                                                  "Delete",
+                                                  style: TextStyle(
+                                                      color: Colors.blue),
+                                                ),
                                               ),
                                             ),
-                                          ),
-                                        if (isEditable)
-                                          PopupMenuItem(
-                                            child: TextButton(
-                                              onPressed: () {
-                                                _showMyDialog(
-                                                  familyList[index]
-                                                      .userId
-                                                      .toString(),
-                                                  familyList[index]
-                                                      .uniqueUserID
-                                                      .toString(),
-                                                );
-                                              },
-                                              child: Text(
-                                                "Delete",
-                                                style: TextStyle(
-                                                    color: Colors.blue),
-                                              ),
-                                            ),
-                                          ),
-                                      ],
+                                        ],
+                                      ),
                                     ),
                                   ),
                                 ],
@@ -207,7 +207,10 @@ class _FamilyListState extends State<FamilyList> {
   Future<bool> _onBackPressed() {
     Navigator.pushReplacement(
       context,
-      MaterialPageRoute(builder: (context) => NavBar(index: 0,)),
+      MaterialPageRoute(
+          builder: (context) => NavBar(
+                index: 0,
+              )),
     );
     return Future.value(false);
   }
@@ -236,7 +239,42 @@ class _FamilyListState extends State<FamilyList> {
               onPressed: () {
                 Navigator.pushReplacement(
                   context,
-                  MaterialPageRoute(builder: (context) => NavBar(index:1)),
+                  MaterialPageRoute(builder: (context) => NavBar(index: 1)),
+                );
+                dltMemberService.deleteFamilyMember(userId, uniqueUserID);
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  Future<void> showMyDialog(String userId, String uniqueUserID) async {
+    return showDialog<void>(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          content: SingleChildScrollView(
+            child: Column(
+              children: [
+                Text('Are you sure want to remove'),
+              ],
+            ),
+          ),
+          actions: [
+            TextButton(
+              child: Text('Cancel'),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+            TextButton(
+              child: Text('Confirm'),
+              onPressed: () {
+                Navigator.pushReplacement(
+                  context,
+                  MaterialPageRoute(builder: (context) => NavBar(index: 1)),
                 );
                 dltMemberService.deleteFamilyMember(userId, uniqueUserID);
               },
